Audit high-value refactors; safe dependency updates
- Document 5 ranked refactors (authz, hours/tax, TS migration,
  Knex/build, frontend consolidation) in plans/
- Bump patch/minor deps within current majors; leave React/Vite/
  Polaris/i18next majors alone
- Move mocha/chai to devDependencies; remove unused history
- Align @types/react to v16 with React 16 runtime
- Fix broken shared imports in kitchen auto-print helpers
- Drop unused lodash require from merchants routes

Authored by Cursor Agent
