+++ package.json
@@ -35,61 +35,60 @@
     "node": ">=22.14.0"
   },
   "dependencies": {
-    "@aws-sdk/client-s3": "^3.1075.0",
-    "@aws-sdk/s3-request-presigner": "^3.1075.0",
+    "@aws-sdk/client-s3": "^3.1086.0",
+    "@aws-sdk/s3-request-presigner": "^3.1086.0",
     "@shopify/polaris": "^10.50.1",
-    "@supabase/supabase-js": "^2.108.2",
+    "@supabase/supabase-js": "^2.110.3",
     "camelcase-keys": "^6.2.2",
-    "chai": "^4.5.0",
     "cookie-parser": "~1.4.7",
-    "date-fns": "^4.1.0",
+    "date-fns": "^4.4.0",
     "date-fns-tz": "^3.2.0",
     "debug": "~4.4.3",
     "dotenv": "^17.4.2",
     "ejs": "^3.1.10",
     "express": "^5.2.1",
-    "history": "^5.3.0",
     "http-errors": "~2.0.1",
     "i18next": "^25.10.10",
-    "knex": "^3.2.10",
+    "knex": "^3.3.0",
     "lodash": "^4.18.1",
-    "mocha": "^11.7.6",
     "morgan": "~1.11.0",
-    "pg": "^8.21.0",
+    "pg": "^8.22.0",
     "react": "^16.14.0",
     "react-bootstrap": "^1.6.8",
     "react-dom": "^16.14.0",
     "react-i18next": "^16.6.6",
     "react-router-dom": "^5.3.4",
-    "sharp": "^0.35.2",
+    "sharp": "^0.35.3",
     "sqlite3": "^5.1.7",
-    "tsx": "^4.22.4",
-    "uuid": "^14.0.0"
+    "tsx": "^4.23.1",
+    "uuid": "^14.0.1"
   },
   "devDependencies": {
     "@modyfi/vite-plugin-yaml": "^1.1.1",
-    "@tailwindcss/vite": "^4.3.0",
+    "@tailwindcss/vite": "^4.3.2",
     "@types/cookie-parser": "^1.4.10",
     "@types/debug": "^4.1.13",
     "@types/express": "^5.0.6",
     "@types/lodash": "^4.17.24",
     "@types/morgan": "^1.9.10",
-    "@types/node": "^24.12.4",
+    "@types/node": "^24.13.3",
     "@types/pg": "^8.20.0",
-    "@types/react": "^19.2.16",
-    "@types/react-dom": "^19.2.3",
+    "@types/react": "^16.14.70",
+    "@types/react-dom": "^16.9.25",
     "@types/react-router-dom": "^5.3.3",
     "@types/supertest": "^7.2.0",
     "@vitejs/plugin-react": "^5.2.0",
-    "autoprefixer": "^10.5.0",
-    "concurrently": "^9.2.1",
+    "autoprefixer": "^10.5.2",
+    "chai": "^4.5.0",
+    "concurrently": "^9.2.4",
+    "mocha": "^11.7.6",
     "nodemon": "^3.1.14",
-    "playwright": "^1.60.0",
-    "postcss": "^8.5.15",
+    "playwright": "^1.61.1",
+    "postcss": "^8.5.19",
     "supertest": "^7.2.2",
-    "tailwindcss": "^4.3.0",
+    "tailwindcss": "^4.3.2",
     "typescript": "^5.9.3",
-    "vite": "^7.3.5"
+    "vite": "^7.3.6"
   },
   "packageManager": "pnpm@10.34.1",
   "pnpm": {
